Moving utilities for the component-library app, so they don't show up in the list of cypress tests to run
