+++ test/e2e/integration/component-library/summary-of-previous-task.ts
@@ -1,10 +1,10 @@
-import { fillInAddressAndVerify } from 'test/e2e/integration/component-library/utils/fillAddressAndVerify';
-import { fillInInputAndVerify } from 'test/e2e/integration/component-library/utils/inputAndVerify';
+import { AppFrontend } from 'test/e2e/pageobjects/app-frontend';
+import { fillInAddressAndVerify } from 'test/e2e/support/apps/component-library/fillAddressAndVerify';
+import { fillInInputAndVerify } from 'test/e2e/support/apps/component-library/inputAndVerify';
 import {
   uploadFileAndVerify,
   uploadFileWithTagAndVerify,
-} from 'test/e2e/integration/component-library/utils/uploadFileAndVerify';
-import { AppFrontend } from 'test/e2e/pageobjects/app-frontend';
+} from 'test/e2e/support/apps/component-library/uploadFileAndVerify';
 
 const appFrontend = new AppFrontend();
